Helpers refactor
