+++ react/components/constants.js
@@ -1,0 +1,34 @@
+export const CONSTANTS = {
+  ICON: {
+    CLASS: 'c-on-base',
+    ANIMATION: 'animated faster zoomIn',
+  },
+  LABEL: {
+    CLASS: 'c-on-base',
+    ANIMATION: 'animated faster fadeIn',
+  },
+  LOGO: {
+    MOBILE: {
+      WIDTH: 90,
+      HEIGHT: 40,
+    },
+    DESKTOP: {
+      WIDTH: 132,
+      HEIGHT: 40,
+    },
+  },
+  SCROLL_ANIMATION: {
+    ANCHOR: 100,
+    REVERSE: false,
+    SPEED: 'faster',
+    ON_SCROLL_DOWN: 'slideOutUp',
+    ON_SCROLL_UP: 'slideInDown',
+  },
+  SPACER: {
+    MOBILE: 96,
+    DESKTOP: 160,
+  },
+  RESIZE_BREAKPOINTS: {
+    MOBILE: 640,
+  },
+}
